Move inventory groups tool back to ansible module
