+++ agent/tools/ansible.py
@@ -1,3 +1,4 @@
+import configparser
 import logging
 import os
 import pathlib
@@ -15,6 +16,7 @@ logger = logging.getLogger(__name__)
 
 ANSIBLE_TMP_DIR: Final[pathlib.Path] = pathlib.Path(".ansible/tmp")
 PLAYBOOKS_DIR: Final[pathlib.Path] = pathlib.Path("ansible/playbooks")
+INVENTORY_PATH: Final[pathlib.Path] = pathlib.Path("ansible/inventory.ini")
 
 
 class AnsiblePlaybookMetadata(BaseModel):
@@ -147,6 +149,14 @@ def get_ansible_playbook_by_name(name: str) -> AnsiblePlaybookRegistryEntry:
     if len(matches) > 1:
         raise ValueError(f"Playbook name is ambiguous in the registry: {name}")
     return matches[0]
+
+
+@tool
+def get_ansible_inventory_groups() -> list[str]:
+    """Return the supported top-level inventory groups from ansible/inventory.ini."""
+    parser = configparser.ConfigParser(allow_no_value=True)
+    parser.read(INVENTORY_PATH, encoding="utf-8")
+    return sorted(section for section in parser.sections() if ":" not in section)
 
 
 @tool
